Add and rename variables and columns of SAR flags and bank IDs

+++ scripts/clean_logs.sh
@@ -8,6 +8,8 @@ rm -f ${OUTPUT}/*.csv
 rm -f ${OUTPUT}/*.png
 rm -f ${OUTPUT}/*.gif
 
+if [ -d ${TMP} ]; then
 rm -rf ${TMP:?}/*/
 rm -f ${TMP}/*.csv
+fi
